Updated and debugged Docker setup
The DockerFile would not build and there seemed to be errors in the
instructions for building and running localing using Docker. Both
the DockerFile and the Instructions have been corrected.  The
build is based on the cu12-0.11.0 Cuda-Q image and minimal packages
are installed afterwards.  Furthermore, all labs are copied to
the container and their file structure is preserved.  As it is,
the setup runs on a an Ubuntu24.04 based system. Further testing
on Windows and MacOS environments are needed.